+++ castor/lib/core/src/main/scala/io/iohk/atala/castor/core/model/did/w3c/W3CModelHelper.scala
@@ -63,6 +63,14 @@ private[castor] trait W3CModelHelper {
             x = x.toStringNoPadding,
             y = y.toStringNoPadding
           )
+        case pk: PublicKeyData.ECCompressedKeyData =>
+          val uncompressed = pk.toUncompressedKeyData
+          PublicKeyJwk(
+            kty = "EC",
+            crv = uncompressed.crv.name,
+            x = uncompressed.x.toStringNoPadding,
+            y = uncompressed.y.toStringNoPadding
+          )
       }
     )
   }